pyproject.toml
[tool.poetry]
name = "pythonproject2"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.15.2"
mysqlclient = "^2.2.6"
daphne = "^4.1.2"
channels = "^4.2.0"
rest-framework-simplejwt = "^0.0.2"
djangorestframework-simplejwt = "^5.3.1"
djangochannelsrestframework = "^1.2.0"
requests = "^2.32.3"
celery = "^5.4.0"
redis = "4.3.5"
django-celery-beat = "^2.7.0"


[tool.poetry.group.dev.dependencies]
django-dotenv = "^1.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
